// src/app/models/retail/retailProductReportResponse.model.ts
export class IRetailProductReportResponse {
    avg_brand: string;
    avg_price: string;
    avg_quality: string;
    company_count: string;
    index_max: string;
    index_min: string;
    local_market_level: string;
    local_market_size: string;
    local_price: string;
    local_quality: string;
    name: string;
    shop_count: string;
    symbol: string;
}